feat(prefab): add device details about headset and controllers
The new DeviceDetailsRecord has been extended into the
OVRInputDetailsRecord and the OVRPluginDetailsRecord that provide
information about an Oculus controller or headset device and this
is now stored on the prefab so it is easily available at runtime.

The DominantControllerObserver has also been added to the prefab to
make it easier to determine the current dominant controller.