feat(DeviceDetails): add passthrough camera details
The DeviceDetailsRecord now supports passthrough camera options and
the Oculus Integration SDK supports a PassThrough layer for the Headset
cameras.

This adds the ability to access the passthrough SDK via a generic
mechanism on the OVRPluginDetailsRecord, which is used for the
DeviceDetailsRecord for the headset.

The OVRInputDetailsRecord, which is used for controllers just
implements a default false output as no controllers support passthrough
as of yet.