fix(Tracking): ensure controller angular velocity is correct
The controller angular velocity was flipped when deploying to quest.
This has now been fixed by flipping the reported angular velocity.

The way in which the relative to body is calculated is also different
based on a Quest or Rift headset (for some random reason?!) so there
is a new enum that allows the selection of whether to use the
TrackedGameObject or the RelativeTo property for the offset multiplier
and this will then need to be set based on whether it is a rift or
quest headset. This can be done by rules and pattern matching so
there won't be any automatic way in case this doesn't catch all
scenarios that may come with future headsets.